requirements.txt
streamlit 
python-dotenv
langchain
langchain_groq
langchain_core
pydantic